+++ source/java/com/pb/mtctm2/abm/ctramp/TourModeChoiceDMU.java
@@ -287,12 +287,12 @@ public class TourModeChoiceDMU
     }
     
     //HH-level auto operating cost [Edit bmp @ July 2019]
-    public double getFuelCost()
+    public float getFuelCost()
     {
     	return hh.getFuelCost();
     }
     
-    public double getMaintainCost()
+    public float getMaintainCost()
     {
         return hh.getMaintainCost();
     }
